fix: address blocking and non-blocking review findings in enhanced init command
Blocking fixes:
- Fix 1: Tests in init-enhanced.test.ts and init-yes-summary.test.ts now run in
  isolated temp dirs instead of chdir'ing to the real repo root; no longer touch
  the tracked CLAUDE.md file
- Fix 2: Reduce init.ts from 837 lines to 83 lines by extracting runInteractive()
  into init-interactive.ts and runNonInteractive() + shared helpers into
  init-runners.ts (all files now under 500 lines)

Non-blocking fixes:
- Fix 3: Remove re-declared HarnessName in init-pipeline.ts; now imported from
  overrides.ts (canonical source) and re-exported for existing consumers
- Fix 4: Export buildCompanionFragment from companion.ts; init-pipeline.ts imports
  both companionSentinel and buildCompanionFragment from there instead of duplicating
- Fix 5: Extend remote-detection regexes to handle ssh:// protocol and
  HTTPS URLs with embedded credentials (user:token@github.com)
- Fix 6: Replace fragile /\{[\s\S]*\}/ JSON extraction regex with lastIndexOf-based
  extraction that correctly finds the top-level JSON object boundary
